Add prometheus handling and logging to tech test.
I was messing around with prometheus metrics and was using the techtest
as a sample app. We may as well add this to the app and maybe someone
will plumb it in.

The task count is not going to be perfect when monitoring from the
application. It's easy to get out of date with multiple copies running.

Some of the gorilla handlers seem like just the defaults we should do
(particularly logging to stdout, and recovery).